Inserting data into dimension tables before facts.
In the previous approach, dimensions were inserted according to the entry of a fact. If the dimensions associated with a fact did not exist in the dimension tables, they were inserted on-the-fly. However, this can cause problems if the analyst wants to check users without shopping carts. But the previous approach can be useful in cases where many facts are generated at the same time. Therefore, the extract_realtime_facts.ktr file was created based on this approach.